tests/uat/pyproject.toml
[project]
name = "my-robot-uat"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "websocket-client>=1.7.0",
    "pytest>=8.0.0",
]
